requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile ../requirements.in
#
backcall==0.1.0           # via ipython
decorator==4.4.0          # via ipython, traitlets
django-ordered-model==3.1.1
django==2.2.1
ipdb==0.12
ipython-genutils==0.2.0   # via traitlets
ipython==7.5.0            # via ipdb
jedi==0.13.3              # via ipython
parso==0.4.0              # via jedi
pexpect==4.7.0            # via ipython
pickleshare==0.7.5        # via ipython
prompt-toolkit==2.0.9     # via ipython
ptyprocess==0.6.0         # via pexpect
pygments==2.3.1           # via ipython
pytz==2019.1              # via django
six==1.12.0               # via prompt-toolkit, traitlets
sqlparse==0.3.0           # via django
traitlets==4.3.2          # via ipython
wcwidth==0.1.7            # via prompt-toolkit
